Refactor NFT functionality: Move from lib to tools directory, improve type safety, use structured parameters, and fix circular dependencies
